common/pyproject.toml
[project]
name = "poweretl-common"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
]

[tool.poetry]
packages = [{include = "poweretl", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
poweretl-defs = { path = "../defs/" }
pytest = "^7.4.0"
pytest-cov = "^7.0.0"

[tool.poetry.dependencies]
poweretl-defs = "*"
dacite = "^1.9.2"
deepmerge = "^2.0"
pyyaml = "^6.0.3"
